Generates passwords on Linux and Andrdoid
On Linux, xdotool and zenity are used to interact with the user and
type the generated password.  On Android, termux-dialog is used to
interact with the user and Android Debug Bridge must be used to type
the password as if the user typed it.

If firefox is present, the actually opened web page is used to guess
the site (prefix) of the password.

All informations used to generate a password (except the master
password itself) is kept on the file system to avoid typing iterations
and length informations on the next run.

OpenSSL is used to generate SHA1 hashes and to encode/decode base64.