Allow configuration of password filter
In order to choose if password must contain only alphanumeric character,
anything or a specific pattern, the user is asked for a filter when
generating a password for the first time.  by defaut, the proposed
regular expression keeps only alphanumeric character.  To disable the
filter, the user may enter "none".