Introduce random salt support
Now, the user has the choice to append a salt to his password.  If so,
the salt (32 random characters) is appended to the string that will be
hashed to become the resulting password and then saved alongside with
the rest of the parameters (length, iterations, filter) to be reloaded
the next time this pasword has to be generated.

Also, the "ask" adapter extension method has been renamed to "input"
and an new adapter extension method called "question" was introduced.

This adapter extension method is actually reusing "input" when there is
no implementation available.

x11/zenity.sh provides an implemetation of "question".